+++ apps/client/src/components/features/mypage/detail/DetailTitle.tsx
